+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Update to [React 0.12.1](https://github.com/facebook/react/releases/tag/v0.12.1)
 * Move to [klayjs-noflo](https://github.com/noflo/klayjs-noflo)
 * Fix Windows build (#192)
+* `React.createFactory` and `displayName` for all React elements
 
 ## 0.3.5 (Durham, 2014 November 14)